service b/requirements.txt
fastapi
uvicorn[standard]
kafka-python
psycopg2-binary
httpx
python-dotenv